refactor(ui): #343 video module + queryKeys (VideoDetails)
- video.ts: fetch/delete/neighbors/frames/fusion, storage day, regen, merge
- JOB_STATUS_POLL_TIMEOUT_MS in client.ts
- queryKeys: video, birdDirectory, speciesSummary
- invalidateLocalSpeciesCaches uses queryKeys
- video.test.ts neighbors params

Made-with: Cursor
