+++ app/ui/src/api/api.tsx
@@ -4,7 +4,6 @@ import {
   SpeciesSummary,
   OverviewData,
   Species,
-  TrackFrame,
 } from '../types';
 import axios from 'axios';
 import { formatLocalTime } from '../util';
@@ -12,7 +11,11 @@ import type { TimeOfDay } from '../utils/timeUtils';
 import type { components } from '../generated/openapi-types';
 import { BASE_API_URL, BASE_URL } from './client';
 
-export { BASE_URL, BASE_API_URL } from './client';
+export {
+  BASE_URL,
+  BASE_API_URL,
+  JOB_STATUS_POLL_TIMEOUT_MS,
+} from './client';
 
 /** Текст ошибки из JSON `{ error: string }` или fallback (для мутаций UI). */
 export function getApiErrorMessage(err: unknown, fallback: string): string {
@@ -27,9 +30,6 @@ export function getApiErrorMessage(err: unknown, fallback: string): string {
   if (err instanceof Error && err.message) return err.message;
   return fallback;
 }
-
-/** Длинный timeout для опроса фоновых job (spectrogram / tracks): дефолт 30s рвёт poll на медленном ответе. */
-export const JOB_STATUS_POLL_TIMEOUT_MS = 120_000;
 
 /**
  * Resolve image URL for display.
@@ -124,112 +124,24 @@ export const fetchRegionComparison = async (): Promise<{
   }
 };
 
-export const fetchVideo = async (id: string) => {
-  const response = await axios.get(`${BASE_API_URL}/videos/${id}`);
-  return response.data;
-};
-
-/** Покадровые bbox — отдельно от GET /videos/:id (лёгкая первая отрисовка страницы). */
-export const fetchVideoDetectionFrames = async (id: string) => {
-  const response = await axios.get(
-    `${BASE_API_URL}/videos/${id}/detection-frames`,
-  );
-  return response.data as {
-    tracks: Array<{
-      id: number | null;
-      species_id: number;
-      start_time: number;
-      end_time: number;
-      frames: TrackFrame[];
-    }>;
-  };
-};
-
-/** Prev/next video IDs for the selected day scope. */
-export type VideoNeighbors = {
-  day_scope: 'utc' | 'local';
-  day_label: string;
-  timezone_offset_minutes: number;
-  cross_day: boolean;
-  previous_id: number | null;
-  next_id: number | null;
-  index: number;
-  total: number;
-};
-
-export const fetchVideoNeighbors = async (
-  id: string,
-): Promise<VideoNeighbors> => {
-  const tzOffset = new Date().getTimezoneOffset();
-  const response = await axios.get(`${BASE_API_URL}/videos/${id}/neighbors`, {
-    params: {
-      day_scope: 'local',
-      tz_offset_minutes: tzOffset,
-      cross_day: 1,
-    },
-  });
-  return response.data;
-};
-
-export type FusionTraceLine = { field: string; value: string };
-export type FusionTraceStep = { stage: string; lines: FusionTraceLine[] };
-export type FusionTraceTrack = {
-  bucket: 'persisted' | 'rejected' | 'accepted';
-  track_id?: number | null;
-  species_name?: string | null;
-  steps: FusionTraceStep[];
-};
-
-export type FusionTracePayload = {
-  available: boolean;
-  video_id?: number;
-  video_path?: string;
-  message?: string;
-  log_created_at?: string | null;
-  trace?: Record<string, unknown> | null;
-  tracks?: FusionTraceTrack[];
-};
-
-export const fetchVideoFusionTrace = async (
-  id: number,
-): Promise<FusionTracePayload> => {
-  const response = await axios.get(
-    `${BASE_API_URL}/videos/${id}/fusion-trace`,
-    {
-      withCredentials: true,
-    },
-  );
-  return response.data;
-};
-
-export const fetchNearestRecordingDay = async (
-  date: string,
-  direction: 'prev' | 'next',
-): Promise<{
-  date: string | null;
-  direction: 'prev' | 'next';
-  found: boolean;
-}> => {
-  const response = await axios.get(
-    `${BASE_API_URL}/storage/nearest-recording-day`,
-    {
-      params: { date, direction },
-    },
-  );
-  return response.data;
-};
-
-/** Delete video recording. Requires contributor or admin access. */
-export const deleteVideo = async (id: number): Promise<void> => {
-  const res = await fetch(`${BASE_API_URL}/videos/${id}`, {
-    method: 'DELETE',
-    credentials: 'include',
-  });
-  if (!res.ok) {
-    const err = await res.json().catch(() => ({ error: res.statusText }));
-    throw new Error(err.error || res.statusText);
-  }
-};
+export {
+  deleteVideo,
+  fetchNearestRecordingDay,
+  fetchVideo,
+  fetchVideoDetectionFrames,
+  fetchVideoFusionTrace,
+  fetchVideoNeighbors,
+  mergeVideoSpecies,
+  regenerateSpectrogramForSingleVideo,
+  regenerateTracksForSingleVideo,
+} from './video';
+export type {
+  FusionTraceLine,
+  FusionTracePayload,
+  FusionTraceStep,
+  FusionTraceTrack,
+  VideoNeighbors,
+} from './video';
 
 /** Export dataset crops as ZIP. Requires settings access. */
 export const exportDataset = async (params?: {
@@ -1626,19 +1538,6 @@ export const fetchTrackRegenerationStatus =
     return response.data;
   };
 
-/** Перегенерация YOLO-треков для одной записи (только админ при двух паролях). */
-export const regenerateTracksForSingleVideo = async (
-  videoId: number,
-  options?: { force?: boolean },
-): Promise<{ message: string; started: boolean; video_id: number }> => {
-  const response = await axios.post(
-    `${BASE_API_URL}/videos/${videoId}/regenerate-tracks`,
-    { force: options?.force === true },
-    { withCredentials: true, timeout: JOB_STATUS_POLL_TIMEOUT_MS },
-  );
-  return response.data;
-};
-
 export interface SpectrogramRegenProgress {
   processed?: number;
   total?: number;
@@ -1666,18 +1565,6 @@ export const fetchSpectrogramRegenerationStatus =
     );
     return response.data;
   };
-
-/** Пересборка спектрограммы для одной записи (только админ при двух паролях). */
-export const regenerateSpectrogramForSingleVideo = async (
-  videoId: number,
-): Promise<{ message: string; started: boolean; video_id: number }> => {
-  const response = await axios.post(
-    `${BASE_API_URL}/videos/${videoId}/regenerate-spectrogram`,
-    {},
-    { withCredentials: true, timeout: JOB_STATUS_POLL_TIMEOUT_MS },
-  );
-  return response.data;
-};
 
 export interface MigrationCalendarData {
   species: Array<{
@@ -1956,17 +1843,5 @@ export const fetchRecentCorrections = async (
     params: { limit },
     withCredentials: true,
   });
-  return response.data;
-};
-
-export const mergeVideoSpecies = async (
-  videoId: string | number,
-  speciesId: number,
-): Promise<{ message: string; species_id: number; updated_count: number }> => {
-  const response = await axios.post(
-    `${BASE_API_URL}/videos/${videoId}/merge-species`,
-    { species_id: speciesId },
-    { withCredentials: true },
-  );
   return response.data;
 };
